third_party/blink/web_tests/external/wpt/svg/linking/scripted/a.type-getter-01-expected.txt

This is a testharness.js-based test.
[FAIL] Test anchor's type getter when type attribute missing
  assert_equals: expected (string) "" but got (undefined) undefined
Harness: the test ran to completion.
